review_saas/requirements.txt
fastapi==0.115.0
    uvicorn[standard]==0.30.6
    jinja2==3.1.4
    SQLAlchemy==2.0.31
    pydantic==2.9.2
    pydantic-settings==2.4.0
    python-dotenv==1.0.1
    email-validator==2.2.0
    passlib[bcrypt]==1.7.4
    python-jose[cryptography]==3.3.0
    reportlab==4.1.0
    httpx==0.27.2
    pandas==2.2.2
    openpyxl==3.1.5
    apscheduler==3.10.4
    psycopg2-binary==2.9.9
    authlib==1.3.1
    pyotp==2.9.0
    bleach==6.1.0
    matplotlib==3.8.4
    langdetect==1.0.9